+++ doc/CHANGELOG.md
@@ -7,7 +7,13 @@
         hash of the response parameters. This is commonly termed as rpHash.
       - Added option **-S**, **\--session** to specify to specify an auxiliary
         session for auditing and or encryption/decryption of the parameters.
-
+      - Added option **-n**, **\--name** to specify the name of the nvindex in
+        hex bytes. This is used when cpHash is calculated without dispatching
+        the TPM2_NV_Write command to the TPM.
+      - Added option **\--with-policydelete** when calculating cpHash with
+        **\--tcti=none** there is no way to know if the attribute
+        TPMA_NV_POLICYDELETE has been set from the NV index name alone.
+    
   * tpm2_nvreadlock:
       - Added option **\--rphash**=_FILE_ to specify file path to record the
         hash of the response parameters. This is commonly termed as rpHash.